Use bigint for all foreign key columns in migrations
Updated all migrations to use bigint type for foreign key columns to
match the bigint primary keys created by Rails. This fixes foreign key
type mismatch errors when deploying to MySQL production.

Changes:
- Updated t.references calls to include `type: :bigint`
- Changed t.integer foreign key columns to t.bigint
- Applied to all join tables and relationship tables

Affected migrations:
- create_tocs: contributor_id, reviewer_id
- create_people_works: person_id, work_id
- create_work_relationships: work1_id, work2_id, creator_id, reviewer_id
- create_realizations: realizer_id, expression_id
- create_reifications: work_id, expression_id
- create_expression_relationships: exp1_id, exp2_id, creator_id, reviewer_id
- create_embodiments: expression_id, manifestation_id
- add_manifestation_id_to_toc: manifestation_id
- create_aboutnesses: embodiment_id
- add_verification_to_aboutnesses: contributor_id, reviewer_id
- create_people_tocs: person_id, toc_id

This ensures foreign keys match their referenced primary keys, preventing
deployment errors on MySQL where type mismatches are enforced.

+++ db/migrate/20150713064700_create_tocs.rb
@@ -4,8 +4,8 @@ class CreateTocs < ActiveRecord::Migration[7.1]
       t.string :book_uri
       t.text :toc_body
       t.string :status
-      t.integer :contributor_id
-      t.integer :reviewer_id
+      t.bigint :contributor_id
+      t.bigint :reviewer_id
       t.text :comments
 
       t.timestamps null: false
