+++ db/migrate/20251126194546_create_aboutnesses.rb
@@ -1,7 +1,7 @@
 class CreateAboutnesses < ActiveRecord::Migration[7.2]
   def change
     create_table :aboutnesses do |t|
-      t.integer :embodiment_id
+      t.bigint :embodiment_id
       t.string :subject_heading_uri
       t.string :source_name
       t.string :subject_heading_label
